+++ docs/source/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fix
+
+- Bug: The initial individual was not printed properly.
+
 ## [3.2.0] - 2022.11.28
 
 ### Fix